ui: display user-friendly exit messages before exec()
There is a duration when we have exec()'ed the real init but the system
display services have not started yet. This results in the content of
our last displayed framebuffer to be stuck until the boot animation
shows up.

This is not a great user experience. Instead, let's display a
"Booting..." message here.