boot-target: test fallocate support by creating a small temp file
Later we would like to be able to distinguish between fallocate errors
due to it being unsupported by the fs and errors due to fs lacking
enough space. In the original approach, any failure results in falling
back to `dd`, which is not what we want to happen when there is
insufficient space left on the device.