import minimal fastbootd implementation from recovery
The fastbootd is copied from recovery but with all commands stripped out
for now. Dependency on Android HIDL is also removed.

I plan to implement fastboot flash support for adding images to
boot-diverter. This way, the user will not need to simg2img and copy the
image to sdcard manually.