paralloid_ui: do not run as pid 1, and keep drm fd open
On some devices, e.g. Motorola G50, calling close() on the master drm fd
would result in the touchscreen being reset, losing some volatile state
in the controller that is essential for the firmware loading process. To
remedy this, we need to drop DRM master while holding the fd itself for
a while before the main system boots.

To do this, we must run paralloid_ ui as non-pid-1, because otherwise
Android's init will close all open fds when it is exec()'ed.

We still keep the init script split into `init` and `boot-target`,
because this way the code still seems better organized.

Doing so also enables the possibility of starting paralloid_ui earlier
in the boot process so that it does not seem stuck while waiting for the
externel SD controller to show up. This would be a great improvement to
the user experience.