minfastbootd: add error message for metadata-encrypted userdata
Partially addresses issue #11.

We cannot handle this case for now. In the future, we can implement an
option to allow Paralloid to completely take over userdata and disable
the ability to boot to system_orig.

TODO: we should also print out friendly error message for the case where
userdata has just been formatted and we need the system to boot first to
set it up properly.