force gatekeeperd to apply uid offset
This is also used in Google's DSU implementation. We replace the GSI
prop with our own prop.

This offset must be applied to avoid conflict between the host image and
other images because otherwise they will all try to use the same
Gatekeeper slots. However, currently the offset is hardcoded, meaning
that there is still conflict between different Paralloid images even
after this patch.

In addition, Weaver slots are not handled by this patch. I do not have a
device that has a Weaver HAL at the moment. However, since Weaver HALs
are optional, we could be able to just disable the Weaver HAL to work
around this.