add unique uid offset in gatekeeperd
...so that different ROMs won't interfere with each other when loading
lockscreen secrets to Gatekeeper HAL. This is similar to Google's DSU
implementation, but ours can support multiple different offsets, instead
of just a hardcoded one.

Since the offset only needs to be set when passed to the Gatekeeper HAL,
we install a hook on libhwbinder methods to intercept the Parcel sent to
the HAL. At least by doing so, we do not need to force software
Gatekeeper or Keymaster.

The inject libraries are put in a separate subdirectory in the native/
directory because of the amount of shared compilation options in our
original native binaries. If we were to merge the two directories, we
would have to duplicate a lot of compilation options since they can now
no longer be set globally in Application.mk.

However, doing this still does not solve all the issues with Gatekeeper.
When booting to the internal target, if the userdata was newly
formatted, gatekeeperd will clear the state of Gatekeeper HAL, which
breaks unlocking in all other Paralloid boot targets. We need to find an
elegant way to solve this.