gatekeeperd_hook: supress Gatekeeper state clearing with the coldboot marker
This is way more elegant than doing a `sed` on the binary, which may not
even work in every case since the property is only for DSU GSIs.

By creating the .coldboot marker before gatekeeperd actually starts
running, we can make sure that it never clears the Gatekeeper HAL state.
As an added bonus, we don't even need to worry about the uid being
already shifted due to the GSI prop anymore.