init: apply gatekeeperd hook to internal target as well
Prevent gatekeeperd in system_orig to clear the state of Gatekeeper when
the main userdata is deleted / formatted.

Note that on some devices with "security chips", the mere action of
formatting userdata itself may have triggered a Gatekeeper state reset.
To work around this, we will need to either modify the recovery, or
implement our own factory reset that does not deal with the security
chip, or simply implement the ability to pivot_root to a custom recovery
environment.

This still hasn't addressed potential issues with devices that have the
Weaver HAL. We need a separate hook to remove the Weaver HAL from the
view of app_process on non-internal targets. (There is no conflict here
because when Weaver is removed, non-internal targets will fall back to
storing secrets in Gatekeeper, which we have already handled. The
internal target can keep storing data in Weaver, because nobody else
will need to access Weaver.)