inject: introduce hook for app_process{32,64} to block Weaver HAL
This needs less code than parsing and editing vintf fragments (we cannot
just delete everything related to Weaver, since some Weaver HALs also
act as other HALs).

Note that LD_PRELOAD is not applicable to app_process due to the
AT_SECURE flag. Instead, we introduce `patchelf` to inject the library
directly to the top of the dependency list of app_process. In addition,
linker config has to be edited to accomodate for this (SELinux policy
not written yet).

The patchelf binary added in this commit is version 0.9 as newer
versions have the regression of corrupting some Android binaries. I did
not use a submodule because I intend to upgrade patchelf with a fix
later.

TODO: Maybe we should make all injections use the patchelf approach
instead of LD_PRELOAD; we also need to make sure linker config editing
works on pre-R (not sure if /linkerconfig existed before R).